Move BannerAd component to v2/components
