+++ common/v2/components/BannerAd/BannerAd.tsx
