+++ common/v2/components/BannerAd/constants.ts
