+++ common/v2/components/BannerAd/index.ts
